Base step for "login" system.
Added in two more URL methods for testing the HTTP command, the /login/
command will take a user and password, unencrypted for now, and if it
matches that user id's password, will "log them in". the /check command
will see if anyone is currently "logged in".

Possible error is that only one person at a time can check stuff, due to
the fact that on this server applet thing there is only one
"currentUser" but I'm looking in to how to work around that.