Added in the two webcontroller methods
1. Adds the user to his/her parking lot, which gets added to the
parkingLot manager.
2. Returns a list of users from a specific lot.

The names that I used for the parking lots are listed in the Fill()
command, located at  /fill for the url. Let me know if the functions
dont work in ajax.

I had to change the user class a whole lot, so as a result, some of the
other classes had to be changed. Hesham's as well as my Junit tests
needed to be mostly commented out, just because I changed the way users
are made, and what info it holds. Let me know if anything else doesn't
work and we can figure it out.